Resolved null error with Ingredient and Step lists not being initialized when sending parcelable to detail activity

+++ app/src/main/java/model/Ingredient.java
@@ -37,7 +37,7 @@ public class Ingredient implements Parcelable
             ;
 
     protected Ingredient(Parcel in) {
-        this.quantity = ((long) in.readValue((double.class.getClassLoader())));
+        this.quantity = ((double) in.readValue((double.class.getClassLoader())));
         this.measure = ((String) in.readValue((String.class.getClassLoader())));
         this.ingredient = ((String) in.readValue((String.class.getClassLoader())));
     }
